Block Prefabs and Color Changing
Changed block prefabs to use the colored border sprite on the parent
object with the collider.
Changed block prefabs to have TextMesh as their first child object.
Changed color change script to always have 1 for value, and saturation
to not drop below .5f.
Changed color change script so it also changes the text mesh color to
the same color as the edge color.