Android Build
Still trying to fix Android app. Works fine in editor or as a built exe
for PC. Touch controls do work on android.

Main bugs are the coroutine for for spawning a bumper if the balls get
stuck in a horizontal pattern. Other bug is rounds resetting while balls
are still active and blocks overlapping the previous layer as two rounds
are sometimes started simultaneously.